refactored the whole parsing file
